+++ src/test/java/test/MainPage.java
@@ -1,6 +1,7 @@
-package framework.pageObject;
+package test;
 
 import org.openqa.selenium.By;
+import pageObject.BasePage;
 
 public class MainPage extends BasePage {
     private static By uniqueElement = By.xpath("//a[@class='enter']");
